feat: add textarea for azure provider
